# Backend/api-gateway-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for Pinz mobile client",
        "title": "Pinz API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "pinz.website",
    "basePath": "/",
    "paths": {
...
        "/api/v1/profile/device-tokens": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers an APNS device token for the authenticated user. Idempotent: повторная регистрация того же токена допустима (обновится owner).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Register APNS device token",
                "parameters": [
                    {
                        "description": "APNS token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.DeviceTokenRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.DeviceTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes an APNS device token for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Unregister APNS device token",
                "parameters": [
                    {
                        "description": "APNS token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.DeviceTokenUnregisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.DeviceTokenUnregisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pinz_backend_api-gateway-service_internal_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pinz_backend_api-gateway-service_internal_responses.DeviceTokenRegisterRequest": {
            "type": "object",
            "properties": {
                "apns_token": {
                    "type": "string"
                }
            }
        },
        "pinz_backend_api-gateway-service_internal_responses.DeviceTokenResponse": {
            "type": "object",
            "properties": {
                "token_id": {
                    "type": "string"
                }
            }
        },
        "pinz_backend_api-gateway-service_internal_responses.DeviceTokenUnregisterRequest": {
            "type": "object",
            "properties": {
                "apns_token": {
                    "type": "string"
                }
            }
        },
        "pinz_backend_api-gateway-service_internal_responses.DeviceTokenUnregisterResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "pinz_backend_api-gateway-service_internal_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}